Fix data types for memory allocations. Thanks for Dr. Bill Barth from TACC for the patches.
git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@3001 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/mpe2/src/graphics/include/mpetools.h
@@ -18,10 +18,10 @@
 #include <stdlib.h>
 #endif
 
-#define MALLOC(a)    malloc((unsigned)(a))
+#define MALLOC(a)    malloc((size_t)(a))
 #define FREE(a)      free((char *)(a))
-#define CALLOC(a,b)    calloc((unsigned)(a),(unsigned)(b))
-#define REALLOC(a,b)   realloc(a,(unsigned)(b))
+#define CALLOC(a,b)    calloc((size_t)(a),(size_t)(b))
+#define REALLOC(a,b)   realloc(a,(size_t)(b))
 
 #define NEW(a)    (a *)MALLOC(sizeof(a))
